kinda stable but some layouts are off
